requirements.txt
python-dotenv>=1.0.0
pandas>=2.0.0
matplotlib>=3.8.0
langchain>=0.2.0
langchain-openai>=0.1.0
openai>=1.35.0
